build(creative-agent): split into server/client entries
Vite externalizes node:path in the browser, and the slash refactor
in dd91762 widened the webui client's import graph for
@agentchan/creative-agent enough to reach tools/script.ts and
skill-content.ts — both of which `import { ... } from "node:path"`
at the top level. Vite dev evaluates ESM modules wholesale and
ignores sideEffects hints, so neither tree-shaking nor a flag could
fix it; the client booted to a blank page.

Resolution: a dedicated `client.ts` entry that exports only
browser-safe bindings — slash module, isSkillContentBlock, session
tree algorithms, type-only re-exports. Vite tree-shakes by module
not by binding, so isSkillContentBlock and buildSkillContent had to
be split into separate files first (the former is pure, the latter
keeps node:path).

- skill-content.ts → skill-content-detect.ts + skill-content-build.ts
- New src/client.ts (0.39 KB bundled). Verified pi-agent-core/pi-ai
  are isomorphic (zero `from "node:*"` in their dist) so future
  widening to value re-exports is safe.
- tsdown two-entry build, package.json gains "./client" sub-export
  and "sideEffects": false (audited safe; helps prod tree-shake even
  though dev mode ignores it).
- Six webui client files migrated from "@agentchan/creative-agent"
  to "@agentchan/creative-agent/client".

Verified end-to-end via slash dogfood — A through G all PASS
(autocomplete popup, source labels, always-active filter, local
dispatch, skill expansion, first-message auto-inject without
duplication, unknown-slash passthrough). See dogfood-output/report.md.
