+++ apps/webui/src/client/entities/skill/skill.types.ts
@@ -1,1 +1,1 @@
-export type { SkillMetadata } from "@agentchan/creative-agent";
+export type { SkillMetadata } from "@agentchan/creative-agent/client";
